Remove the nodeupdate compatibility check in its own package for re-use
