+++ api/pkg/handler/v1/common/version_skew.go
@@ -2,7 +2,6 @@ package common
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"strings"
 
@@ -10,49 +9,11 @@ import (
 
 	kubermaticapiv1 "github.com/kubermatic/kubermatic/api/pkg/crd/kubermatic/v1"
 	"github.com/kubermatic/kubermatic/api/pkg/provider"
+	"github.com/kubermatic/kubermatic/api/pkg/validation/nodeupdate"
 
 	clusterv1alpha1 "sigs.k8s.io/cluster-api/pkg/apis/cluster/v1alpha1"
 	ctrlruntimeclient "sigs.k8s.io/controller-runtime/pkg/client"
 )
-
-// ErrVersionSkew denotes an error condition where a given kubelet/controlplane version pair is not supported
-type ErrVersionSkew struct {
-	ControlPlane *semver.Version
-	Kubelet      *semver.Version
-}
-
-// Error returns a string representation of the error
-func (e ErrVersionSkew) Error() string {
-	return fmt.Sprintf("kubelet version %s is not compatible with control plane version %s", e.Kubelet, e.ControlPlane)
-}
-
-var _ error = ErrVersionSkew{}
-
-// EnsureVersionCompatible checks whether the given kubelet version
-// is deemed compatible with the given version of the control plane.
-func EnsureVersionCompatible(controlPlane *semver.Version, kubelet *semver.Version) error {
-	if controlPlane == nil {
-		return errors.New("ensureVersionCompatible: controlPlane is nil")
-	}
-
-	if kubelet == nil {
-		return errors.New("ensureVersionCompatible: kubelet is nil")
-	}
-
-	// Kubelet must be the same major version and no more than 2 minor versions behind the control plane.
-	// https://kubernetes.io/docs/setup/version-skew-policy/
-	// https://github.com/kubernetes/website/blob/076efdf364651859553681a75f60c957de729023/content/en/docs/setup/version-skew-policy.md
-	compatible := kubelet.Major() == controlPlane.Major() && kubelet.Minor() >= (controlPlane.Minor()-2) && kubelet.Minor() <= controlPlane.Minor()
-
-	if !compatible {
-		return ErrVersionSkew{
-			ControlPlane: controlPlane,
-			Kubelet:      kubelet,
-		}
-	}
-
-	return nil
-}
 
 // CheckClusterVersionSkew returns a list of machines and/or machine deployments
 // that are running kubelet at a version incompatible with the cluster's control plane.
@@ -78,9 +39,9 @@ func CheckClusterVersionSkew(ctx context.Context, userInfo *provider.UserInfo, c
 			return nil, fmt.Errorf("failed to parse kubelet version: %v", parseErr)
 		}
 
-		if err = EnsureVersionCompatible(clusterVersion, kubeletVersion); err != nil {
+		if err = nodeupdate.EnsureVersionCompatible(clusterVersion, kubeletVersion); err != nil {
 			// errVersionSkew says it's incompatible
-			if _, ok := err.(ErrVersionSkew); ok {
+			if _, ok := err.(nodeupdate.ErrVersionSkew); ok {
 				incompatibleVersionsSet[kubeletVersion.String()] = true
 				continue
 			}
